Fix Drop Shadow filter to work with partial update rects
1) Fix Drop Shadow LS filter to regenerate partial rect only
2) Fix needRect and changeRect methods
3) Indroduce KisLayerStyleFilter which has a special method for
   cross-device compositioning. It is needed because the source
   of a layer style filter should keep unchanged when processing.

+++ krita/libpsd/psd.cpp
@@ -204,8 +204,8 @@ QPoint psd_layer_effects_drop_shadow::calculateOffset(const psd_layer_effects_co
     qint32 angle = this->use_global_light ?
         context->global_angle : this->angle;
 
-    qint32 distance_x = -qRound(this->distance * cos(kisDegreesToRadians(angle)));
-    qint32 distance_y =  qRound(this->distance * sin(kisDegreesToRadians(angle)));
+    qint32 distance_x = -qRound(this->distance * cos(kisDegreesToRadians(qreal(angle))));
+    qint32 distance_y =  qRound(this->distance * sin(kisDegreesToRadians(qreal(angle))));
 
     return QPoint(distance_x, distance_y);
 }
